requirements_v5.txt
google-generativeai
python-dotenv
rich
Pillow
pillow-lut
chromadb